Add end-to-end tests for organizer restrictions, score entry, and visitor display
- Implement organizer restrictions tests for teams, games, and dashboard stats in `organizer-restrictions.spec.ts`.
- Create `DashboardPage`, `GamesPage`, `PoulesPage`, `TeamsPage`, and `TournamentBasePage` classes for page object model structure.
- Add score entry and standings tests in `score-standings.spec.ts` to validate score display, entry, editing, and YAML export.
- Develop visitor display tests in `visitor-display.spec.ts`, `visitor-games.spec.ts`, and `visitor-poules.spec.ts` to ensure correct visibility and functionality for public users.
- Introduce fixtures for testing with predefined tournament data.
